fixed the characterCustomization
fully works with the rest of the project, I could fix it up to use the first row of tiled images and the animation when displaying the images but for now it is just the static images.

I could also fix up the buttons to match the start menu for continuity as well as changing the background (if the team wants).